KL-149/feat: add ProductService interface
